Fix clang-tidy warnings

+++ src/file-hash.h
@@ -11,7 +11,7 @@
 #include "str.h"
 
 /** The amount of bytes required to store a files hash. */
-#define FILE_HASH_SIZE 20
+#define FILE_HASH_SIZE ((size_t)20)
 
 extern void fileHash(String path, struct stat stats, uint8_t *hash);
